Refactor to place new fetcher in az_core.
