+++ modules/custom/az_core/az_core.info.yml
@@ -14,6 +14,7 @@ dependencies:
   - drupal:taxonomy
   - drupal:telephone
   - drupal:text
+  - guzzle_cache:guzzle_cache
   - linkit:linkit
   - search_exclude:search_exclude
 configure: az_core.az_settings
