+++ modules/custom/az_core/az_core.services.yml
@@ -19,3 +19,9 @@ services:
       - '@config_update.config_list'
       - '@module_handler'
       - '@logger.channel.az_core'
+  cache.az_migration_http_cache:
+     class: Drupal\Core\Cache\CacheBackendInterface
+     tags:
+       - { name: cache.bin }
+     factory: cache_factory:get
+     arguments: [az_migration_http_cache]
