+++ modules/custom/az_core/config/install/az_core.settings.yml
@@ -3,3 +3,5 @@ monitoring_page:
   path: ''
 enterprise_attributes:
   locked: true
+migrations:
+  http_cached: true
