--- modules/custom/az_migration/az_migration_http_cache/az_migration_http_cache.info.yml
@@ -1,11 +1,0 @@
-name: Quickstart Migration HTTP Cache
-description: 'Implements an HTTP cache for Migrate URL-based requests'
-core_version_requirement: ^9 || ^10
-type: module
-package: 'The University of Arizona - Experimental'
-dependencies:
-  - az_core:az_core
-  - migrate:migrate
-  - migrate_plus:migrate_plus
-  - migrate_tools:migrate_tools
-  - guzzle_cache:guzzle_cache
