--- modules/custom/az_migration/az_migration_http_cache/az_migration_http_cache.module
@@ -1,18 +1,0 @@
-<?php
-
-/**
- * @file
- * Contains az_migration_http_cache.module.
- */
-
-/**
- * Implements hook_migration_plugins_alter().
- */
-function az_migration_http_cache_migration_plugins_alter(array &$migrations) {
-  // Swap in custom implementation for migrations which use the HTTP fetcher.
-  foreach ($migrations as &$migration) {
-    if (!empty($migration['source']['data_fetcher_plugin']) && ($migration['source']['data_fetcher_plugin'] === 'http')) {
-      $migration['source']['data_fetcher_plugin'] = 'az_http_cached';
-    }
-  }
-}
