--- modules/custom/az_migration/az_migration_http_cache/az_migration_http_cache.services.yml
@@ -1,7 +1,0 @@
-services:
-  cache.az_migration_http_cache:
-     class: Drupal\Core\Cache\CacheBackendInterface
-     tags:
-       - { name: cache.bin }
-     factory: cache_factory:get
-     arguments: [az_migration_http_cache]
