+++ modules/custom/az_news/az_news_feeds/config/install/migrate_plus.migration_group.az_news_feeds.yml
@@ -5,6 +5,7 @@ dependencies:
     - az_news_feeds
   enforced:
     module:
+      - az_core
       - az_news_feeds
       - migrate_tools
 id: az_news_feeds
@@ -15,7 +16,7 @@ module: az_news_feeds
 shared_configuration:
   source:
     plugin: url
-    data_fetcher_plugin: http
+    data_fetcher_plugin: az_core_http
     data_parser_plugin: json
     urls: 'https://news.arizona.edu/feed/json/stories/id/all'
     constants:
